Moved projectile bullet code out
