+++ Assets/Code/Scripts/AreaOfEffect.cs
@@ -16,7 +16,6 @@ namespace Gun
         {
             base.Init(stats);
             Reset();
-            countPenetration = false;
         }
 
         // Update is called once per frame
