+++ Assets/Code/Scripts/Bullets/ProjectileBullet.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: df71897e533e8c14997d372a079e868e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
