+++ Assets/EditorObject/TestGun.asset
@@ -29,14 +29,22 @@ MonoBehaviour:
   overHeatPercentPerShot: 5
   coolDownPerSecond: 10
   printDebugState: 0
-  bulletType: 2
+  bulletType: 0
   bulletPenetration: 0
   muzzleVelocity: 9.7
   projectileScale: {x: 3, y: 1, z: 3}
   range: 100
-  isExplosive: 1
+  isExplosive: 0
   isCountDownExplosion: 1
   radius: 5
   force: 12000
   explosionDamage: 25
   secondsBeforeExplode: 1
+  damagePerSecond: 10
+  numPhases: 2
+  phase1:
+    Duration: 0
+    RateOfScaleGrowth: 0
+  phase2:
+    Duration: 0
+    RateOfScaleGrowth: 0
