+++ Assets/Prefabs/Bullets/LMGBullet.prefab
@@ -13,14 +13,14 @@ GameObject:
   - component: {fileID: 8579180907413031318}
   - component: {fileID: 8610469206652896614}
   - component: {fileID: 876733522086653541}
-  - component: {fileID: 5230005726319368862}
+  - component: {fileID: 8930960238914550717}
   m_Layer: 0
   m_Name: LMGBullet
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 1
+  m_IsActive: 0
 --- !u!4 &5790650606467457179
 Transform:
   m_ObjectHideFlags: 0
@@ -113,7 +113,7 @@ Rigidbody:
   m_Interpolate: 0
   m_Constraints: 0
   m_CollisionDetection: 0
---- !u!114 &5230005726319368862
+--- !u!114 &8930960238914550717
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -122,6 +122,8 @@ MonoBehaviour:
   m_GameObject: {fileID: 6709974543186267795}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: fe92ff41072c28e48addf9ce03785a48, type: 3}
+  m_Script: {fileID: 11500000, guid: df71897e533e8c14997d372a079e868e, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  hasFiniteLifetime: 0
+  lifetime: 3.4028235e+38
